[ASTERIXDB-3032][RT] Fix Int32 MsgPackAccessors
- user mode changes: no
- storage format changes: no
- interface changes: no

Details:
- Write an int instead of a byte in MsgPackInt32Accessor

Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/16283
Contrib: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>

+++ asterixdb/asterix-external-data/src/main/java/org/apache/asterix/external/library/msgpack/MsgPackAccessors.java
@@ -124,7 +124,7 @@ public class MsgPackAccessors {
             int s = pointable.getStartOffset();
             int i = AInt32SerializerDeserializer.getInt(b, s + 1);
             out.writeByte(INT32);
-            out.writeByte(i);
+            out.writeInt(i);
             return null;
         }
     }
